Fix KW issue to avoid null pointer
Some pointers returned from functions may be NULL, so fix these
KW issues to aviod some null pointer usage.

+++ p2p.c
@@ -641,6 +641,8 @@ cmd_sta_get_p2p_dev_address(struct sigma_dut *dut, struct sigma_conn *conn,
 	const char *intf = get_param(cmd, "interface");
 	char buf[100], resp[200];
 
+	if (intf == NULL)
+		return -1;
 	start_sta_mode(dut);
 	if (get_wpa_status(intf, "p2p_device_address", buf, sizeof(buf)) < 0) {
 		send_resp(dut, conn, SIGMA_ERROR, NULL);
@@ -1428,7 +1430,7 @@ static enum sigma_cmd_result cmd_sta_p2p_dissolve(struct sigma_dut *dut,
 	struct wfa_cs_p2p_group *grp;
 	char buf[128];
 
-	if (grpid == NULL)
+	if (grpid == NULL || intf == NULL)
 		return -1;
 
 	grp = p2p_group_get(dut, grpid);
@@ -1467,7 +1469,7 @@ cmd_sta_send_p2p_invitation_req(struct sigma_dut *dut, struct sigma_conn *conn,
 	struct wpa_ctrl *ctrl;
 	int res;
 
-	if (devid == NULL || grpid == NULL)
+	if (intf == NULL || devid == NULL || grpid == NULL)
 		return -1;
 
 	if (reinvoke && atoi(reinvoke)) {
@@ -1557,7 +1559,7 @@ cmd_sta_accept_p2p_invitation_req(struct sigma_dut *dut,
 	const char *reinvoke = get_param(cmd, "Reinvoke");
 	char buf[100];
 
-	if (devid == NULL || grpid == NULL)
+	if (intf == NULL || devid == NULL || grpid == NULL)
 		return -1;
 
 	if (reinvoke && atoi(reinvoke)) {
@@ -1602,7 +1604,7 @@ cmd_sta_send_p2p_provision_dis_req(struct sigma_dut *dut,
 	char buf[256];
 	char *method;
 
-	if (conf_method == NULL || devid == NULL)
+	if (intf == NULL || conf_method == NULL || devid == NULL)
 		return -1;
 
 	if (strcasecmp(conf_method, "Display") == 0)
@@ -1669,6 +1671,8 @@ static enum sigma_cmd_result cmd_sta_wps_read_pin(struct sigma_dut *dut,
 	char pin[9], addr[20];
 	char resp[100];
 
+	if (intf == NULL)
+		return -1;
 	if (get_wpa_status(intf, "address", addr, sizeof(addr)) < 0 ||
 	    get_wps_pin_from_mac(dut, addr, pin, sizeof(pin)) < 0) {
 		sigma_dut_print(dut, DUT_MSG_DEBUG,
@@ -1808,6 +1812,8 @@ enum sigma_cmd_result cmd_sta_p2p_reset(struct sigma_dut *dut,
 	struct wfa_cs_p2p_group *grp, *prev;
 	char buf[256];
 
+	if (intf == NULL)
+		return -1;
 #ifdef MIRACAST
 	if (dut->program == PROGRAM_WFD ||
 	    dut->program == PROGRAM_DISPLAYR2)
@@ -2088,7 +2094,7 @@ cmd_sta_send_service_discovery_req(struct sigma_dut *dut,
 	const char *devid = get_param(cmd, "P2PDevID");
 	char buf[128];
 
-	if (devid == NULL)
+	if (intf == NULL || devid == NULL)
 		return -1;
 
 	snprintf(buf, sizeof(buf), "P2P_SERV_DISC_REQ %s 02000001",
@@ -2206,6 +2212,8 @@ static int nfc_write_p2p_select(struct sigma_dut *dut, struct sigma_conn *conn,
 	const char *ifname = get_param(cmd, "Interface");
 	char buf[300];
 
+	if (ifname == NULL)
+		return -1;
 	run_system(dut, "killall wps-nfc.py");
 	run_system(dut, "killall p2p-nfc.py");
 
@@ -2255,6 +2263,8 @@ static int nfc_write_config_token(struct sigma_dut *dut,
 	const char *intf = get_param(cmd, "Interface");
 	char buf[200];
 
+	if (intf == NULL)
+		return -1;
 	run_system(dut, "killall wps-nfc.py");
 	run_system(dut, "killall p2p-nfc.py");
 	unlink("nfc-success");
@@ -2337,6 +2347,8 @@ static int nfc_read_tag(struct sigma_dut *dut,
 	run_system(dut, "killall wps-nfc.py");
 	run_system(dut, "killall p2p-nfc.py");
 
+	if (intf == NULL)
+		return -1;
 	ctrl = open_wpa_mon(intf);
 	if (ctrl == NULL) {
 		sigma_dut_print(dut, DUT_MSG_ERROR, "Failed to open "
@@ -2396,6 +2408,8 @@ static int nfc_wps_read_tag(struct sigma_dut *dut,
 	run_system(dut, "killall wps-nfc.py");
 	run_system(dut, "killall p2p-nfc.py");
 
+	if (intf == NULL)
+		return -1;
 	ctrl = open_wpa_mon(intf);
 	if (ctrl == NULL) {
 		sigma_dut_print(dut, DUT_MSG_ERROR, "Failed to open "
@@ -2535,6 +2549,8 @@ static int nfc_wps_read_passwd(struct sigma_dut *dut,
 	char buf[1000];
 	const char *keymgmt, *cipher;
 
+	if (intf == NULL)
+		return -1;
 	run_system(dut, "killall wps-nfc.py");
 	run_system(dut, "killall p2p-nfc.py");
 
@@ -2664,6 +2680,8 @@ static int nfc_wps_read_config(struct sigma_dut *dut,
 	run_system(dut, "killall wps-nfc.py");
 	run_system(dut, "killall p2p-nfc.py");
 
+	if (intf == NULL)
+		return -1;
 	ctrl = open_wpa_mon(intf);
 	if (ctrl == NULL) {
 		sigma_dut_print(dut, DUT_MSG_ERROR, "Failed to open "
@@ -2702,6 +2720,8 @@ static int nfc_wps_connection_handover(struct sigma_dut *dut,
 	run_system(dut, "killall wps-nfc.py");
 	run_system(dut, "killall p2p-nfc.py");
 
+	if (intf == NULL)
+		return -1;
 	ctrl = open_wpa_mon(intf);
 	if (ctrl == NULL) {
 		sigma_dut_print(dut, DUT_MSG_ERROR, "Failed to open "
@@ -2790,6 +2810,8 @@ static int nfc_p2p_connection_handover(struct sigma_dut *dut,
 	run_system(dut, "killall wps-nfc.py");
 	run_system(dut, "killall p2p-nfc.py");
 
+	if (intf == NULL)
+		return -1;
 	ctrl = open_wpa_mon(intf);
 	if (ctrl == NULL) {
 		sigma_dut_print(dut, DUT_MSG_ERROR, "Failed to open "
@@ -2863,7 +2885,7 @@ static enum sigma_cmd_result cmd_sta_nfc_action(struct sigma_dut *dut,
 	const char *oper_chn = get_param(cmd, "OPER_CHN");
 	char buf[256];
 
-	if (oper == NULL)
+	if (intf == NULL || oper == NULL)
 		return -1;
 
 	if (ssid_param)
